feat: 더미 데이터 제거 및 API 호출로 강의 목록 가져오기 기능 추가 #30
