--- lib/msw/handlers.ts
@@ -1,9 +1,0 @@
-import { http, HttpResponse } from 'msw'
-
-export const handlers = [
-  http.get('/api/test', () => {
-    return HttpResponse.json({
-      message: 'hello from msw',
-    })
-  }),
-]
