+++ lib/msw/handlers/index.ts
@@ -1,0 +1,5 @@
+import { lectureHandlers } from './lecture'
+
+export const handlers = [
+  ...lectureHandlers,
+]
