+++ lib/msw/handlers/lecture/index.ts
@@ -1,0 +1,5 @@
+import { getLectures } from './get-lectures'
+
+export const lectureHandlers = [
+  getLectures,
+]
